create_app moved to root app.py file
