+++ wsgi.py
@@ -1,4 +1,4 @@
-from core.app import app
+from app import app
 
 if __name__ == "__main__":
     app.run()
